Classe moved in correct package
