+++ src/main/java/ch/heigvd/res/chill/domain/ClarFleu/Kwak.java
@@ -1,4 +1,4 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.ClarFleu;
 
 import ch.heigvd.res.chill.domain.IProduct;
